+++ webapp/src/main/kotlin/com/koldyr/genealogy/mapper/FamilyEventConverter.kt
@@ -1,20 +1,20 @@
 package com.koldyr.genealogy.mapper
 
-import com.koldyr.genealogy.model.FamilyEvent
-import com.koldyr.genealogy.persistence.FamilyEventRepository
 import ma.glasnost.orika.MappingContext
 import ma.glasnost.orika.converter.BidirectionalConverter
 import ma.glasnost.orika.metadata.Type
 import org.springframework.http.HttpStatus.*
 import org.springframework.web.server.ResponseStatusException
+import com.koldyr.genealogy.model.FamilyEvent
+import com.koldyr.genealogy.persistence.FamilyEventRepository
 
 /**
  * Description of class FamilyEventConverter
  * @created: 2021-10-08
  */
-class FamilyEventConverter(private val familyEventRepository: FamilyEventRepository) : BidirectionalConverter<Int, FamilyEvent>() {
+class FamilyEventConverter(private val familyEventRepository: FamilyEventRepository) : BidirectionalConverter<Long, FamilyEvent>() {
 
-    override fun convertTo(eventId: Int?, type: Type<FamilyEvent>?, context: MappingContext?): FamilyEvent? {
+    override fun convertTo(eventId: Long?, type: Type<FamilyEvent>?, context: MappingContext?): FamilyEvent? {
         if (eventId == null) {
             return null
         }
@@ -22,7 +22,7 @@ class FamilyEventConverter(private val familyEventRepository: FamilyEventReposit
                 .orElseThrow { ResponseStatusException(NOT_FOUND, "Event with id '$eventId' is not found") }
     }
 
-    override fun convertFrom(event: FamilyEvent?, type: Type<Int>?, context: MappingContext?): Int? {
+    override fun convertFrom(event: FamilyEvent?, type: Type<Long>?, context: MappingContext?): Long? {
         return event?.id
     }
 }
